mcp/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-groq>=0.1.0
langchain-community>=0.0.10
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
pyarrow>=14.0.0
requests>=2.28.0
python-dotenv>=1.0.0
flask>=2.0.0